// src/app/authorization/workflow/insurance/insurance-control-names.enum.ts
export enum InsuranceControlNames {
  IS_INFORMATION_CORRECT = 'isInformationCorrect',
  HAS_INSURANCE = 'hasInsurance',
  WANT_TO_UPLOAD = 'wantToUpload',
  HAS_SECONDARY_INSURANCE = 'hasSecondaryInsurance'
}

export enum InsuranceFormControlNames {
  INSURANCE_PLAN_TYPE = 'insurancePlanType',
  PBM_NAME = 'pbmName',
  PBM_PHONE_NUMBER = 'pbmPhoneNumber',
  POLICYHOLDER_ID_NUMBER = 'policyholderIdNumber',
  GROUP_NUMBER = 'groupNumber',
  BIN_NUMBER = 'binNumber',
  PCN_NUMBER = 'pcnNumber',
  POLICYHOLDER_RELATIONSHIP_TO_PATIENT = 'policyholderRelationshipToPatient',
  POLICYHOLDER_DATE_OF_BIRTH = 'policyholderDateOfBirth',
  POLICYHOLDER_FIRST_NAME = 'policyholderFirstName',
  POLICYHOLDER_LAST_NAME = 'policyholderLastName'
}